refactor validate index
